# Improve performance of batch request worker