@hugecannon 2021-03-26
This PR will need some help from a testing point of view
I couldn't find any tests covering this exact scenario and I'm not too familiar with the testing framework.

I did a manual check on the output of the function and it seemed to be the same given my limited inputs.